Add files via upload
This takes Jack's R code and puts it in Python. It does not contain bootstrapping or added confidence intervals. When not treating the data as time series, you get some coefficients for everything except TLT. Otherwise, all coefficients are driven to 0 regardless of alpha optimization. I believe going forward, that when working with the .csv data, that it should not be treated as time series. This is because the data is from trades and is not given at regular time intervals. In SQL with the 5s data, because it is at normal intervals, treating those data as time series would be more appropriate. When I apply this to the interest rates, (which are definitely time series) I will have to figure that stuff out.